Added some tests showing how to compute other terms involving Hbar with T3 and t4 on-the-fly for ec-CC and CC(P;Q).
